Move SDKConfig into Store.ts
